refactor(sandook): move SQLDelight codegen into its own .db package
The generated query and table classes were emitted into
`xyz.ksharma.krail.sandook` — the same package the hand-written repositories
live in. That makes them impossible to tell apart by any package-based
filter, which is why the Kover exclusion
`packages("xyz.ksharma.krail.sandook.sandook")` only ever matched one file
(`KrailSandookImpl`) while the other 37 generated classes counted as covered
production code.

Two changes are both required, because SQLDelight derives packages from two
different places:

- `packageName` in the sqldelight block moves the database interface and its
  impl. On its own it leaves the queries and table classes where they were.
- The `.sq` files move to `sqldelight/xyz/ksharma/krail/sandook/db/`, because
  a `.sq` file's generated package comes from its directory relative to the
  sqldelight source root.

Everything else here is mechanical: consumers repoint their import at the new
path, and files that declare `package xyz.ksharma.krail.sandook` gain the
import they previously did not need because the generated types shared their
package.

No behaviour change, and no schema change — the migrations directory is
outside the package tree, so `verifyMigrations` is unaffected.
