+++ sandook/src/androidHostTest/kotlin/xyz/ksharma/krail/sandook/SavedTripIdentityMigrationTest.kt
@@ -5,6 +5,7 @@ import kotlinx.coroutines.Dispatchers
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
+import xyz.ksharma.krail.sandook.db.KrailSandook
 import kotlin.test.assertEquals
 import kotlin.test.assertFails
 import kotlin.test.assertNotNull
