+++ sandook/src/androidHostTest/kotlin/xyz/ksharma/krail/sandook/UserLifecycleStoreTest.kt
@@ -6,6 +6,7 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
+import xyz.ksharma.krail.sandook.db.KrailSandook
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
