+++ sandook/src/commonMain/sqldelight/xyz/ksharma/krail/sandook/db/AppPreferences.sq
